move slider code into its own component
