+++ catalogue/webapp/components/SearchForm/SearchForm.js
@@ -10,136 +10,12 @@ import { trackEvent } from '@weco/common/utils/ga';
 import { worksUrl } from '@weco/common/services/catalogue/urls';
 import CatalogueSearchContext from '@weco/common/views/components/CatalogueSearchContext/CatalogueSearchContext';
 import VerticalSpace from '@weco/common/views/components/styled/VerticalSpace';
-import { Slider, Rail, Handles, Tracks, Ticks } from 'react-compound-slider';
+import DateSlider from '@weco/catalogue/components/DateSlider/DateSlider';
 
 type Props = {|
   ariaDescribedBy: string,
   compact: boolean,
 |};
-
-const Tick = ({ tick, count }) => {
-  // your own tick component
-  return (
-    <div>
-      <div
-        style={{
-          position: 'absolute',
-          marginTop: 52,
-          marginLeft: -0.5,
-          width: 1,
-          height: 8,
-          backgroundColor: 'silver',
-          left: `${tick.percent}%`,
-        }}
-      />
-      <div
-        style={{
-          position: 'absolute',
-          marginTop: 60,
-          fontSize: 10,
-          textAlign: 'center',
-          marginLeft: `${-(100 / count) / 2}%`,
-          width: `${100 / count}%`,
-          left: `${tick.percent}%`,
-        }}
-      >
-        {tick.value}
-      </div>
-    </div>
-  );
-};
-
-const KeyboardHandle = ({
-  domain: [min, max],
-  handle: { id, value, percent },
-  disabled,
-  getHandleProps,
-}) => {
-  return (
-    <button
-      role="slider"
-      aria-valuemin={min}
-      aria-valuemax={max}
-      aria-valuenow={value}
-      style={{
-        left: `${percent}%`,
-        position: 'absolute',
-        transform: 'translate(-50%, -50%)',
-        zIndex: 2,
-        width: 24,
-        height: 24,
-        borderRadius: '50%',
-        boxShadow: '1px 1px 1px 1px rgba(0, 0, 0, 0.3)',
-        backgroundColor: disabled ? '#666' : '#ffc400',
-      }}
-      {...getHandleProps(id)}
-    />
-  );
-};
-
-const Handle = ({ handle: { id, value, percent }, getHandleProps }) => {
-  return (
-    <div
-      style={{
-        left: `${percent}%`,
-        position: 'absolute',
-        marginLeft: -15,
-        marginTop: 25,
-        zIndex: 2,
-        width: 30,
-        height: 30,
-        border: 0,
-        textAlign: 'center',
-        cursor: 'pointer',
-        borderRadius: '50%',
-        backgroundColor: '#2C4870',
-        color: '#333',
-      }}
-      {...getHandleProps(id)}
-    >
-      <div style={{ fontFamily: 'Roboto', fontSize: 11, marginTop: -35 }}>
-        {value}
-      </div>
-    </div>
-  );
-};
-
-const Track = ({ source, target, getTrackProps }) => {
-  // your own track component
-  return (
-    <div
-      style={{
-        position: 'absolute',
-        height: 10,
-        zIndex: 1,
-        marginTop: 35,
-        backgroundColor: '#546C91',
-        borderRadius: 5,
-        cursor: 'pointer',
-        left: `${source.percent}%`,
-        width: `${target.percent - source.percent}%`,
-      }}
-      {...getTrackProps()} // this will set up events if you want it to be clickeable (optional)
-    />
-  );
-};
-
-const sliderStyle = {
-  // Give the slider some width
-  position: 'relative',
-  width: '100%',
-  height: 80,
-  border: '1px solid steelblue',
-};
-
-const railStyle = {
-  position: 'absolute',
-  width: '100%',
-  height: 10,
-  marginTop: 35,
-  borderRadius: 5,
-  backgroundColor: '#8B9CB6',
-};
 
 const SearchInputWrapper = styled.div`
   background: ${props => props.theme.colors.white};
@@ -346,69 +222,7 @@ const SearchForm = ({ ariaDescribedBy, compact }: Props) => {
                     </label>
                   </VerticalSpace>
                   <VerticalSpace>
-                    <Slider
-                      rootStyle={sliderStyle}
-                      domain={[0, 2020]}
-                      step={100}
-                      mode={2}
-                      values={[0, 2020]}
-                    >
-                      <div style={railStyle} />
-                      <Rail>
-                        {(
-                          { getRailProps } // adding the rail props sets up events on the rail
-                        ) => <div style={railStyle} {...getRailProps()} />}
-                      </Rail>
-                      <Handles>
-                        {({ handles, getHandleProps }) => (
-                          <div className="slider-handles">
-                            {handles.map(handle => (
-                              <>
-                                <KeyboardHandle
-                                  key={handle.id}
-                                  handle={handle}
-                                  domain={[0, 2020]}
-                                  getHandleProps={getHandleProps}
-                                  disabled={false}
-                                />
-                                <Handle
-                                  key={handle.id}
-                                  handle={handle}
-                                  getHandleProps={getHandleProps}
-                                />
-                              </>
-                            ))}
-                          </div>
-                        )}
-                      </Handles>
-                      <Tracks right={false}>
-                        {({ tracks, getTrackProps }) => (
-                          <div className="slider-tracks">
-                            {tracks.map(({ id, source, target }) => (
-                              <Track
-                                key={id}
-                                source={source}
-                                target={target}
-                                getTrackProps={getTrackProps}
-                              />
-                            ))}
-                          </div>
-                        )}
-                      </Tracks>
-                      <Ticks count={15}>
-                        {({ ticks }) => (
-                          <div className="slider-ticks">
-                            {ticks.map(tick => (
-                              <Tick
-                                key={tick.id}
-                                tick={tick}
-                                count={ticks.length}
-                              />
-                            ))}
-                          </div>
-                        )}
-                      </Ticks>
-                    </Slider>
+                    <DateSlider />
                   </VerticalSpace>
                 </VerticalSpace>
               )}
